KTO-860: Use a normal .ts-file for haku types
- Change HAKULOMAKETYYPPI to enum
- Move HakuFormValues type to a normal ts-file, so HAKULOMAKETYYPPI enum can be used directly.
