+++ src/main/app/src/utils/haku/getFormValuesByHaku.ts
@@ -1,6 +1,7 @@
 import _ from 'lodash/fp';
 import getHakulomakeFieldsValues from '#/src/utils/form/getHakulomakeFieldsValues';
 import { ALKAMISKAUSITYYPPI, TOTEUTUKSEN_AJANKOHTA } from '#/src/constants';
+import { HakuFormValues } from '#/src/types/hakuTypes';
 
 export const alkamiskausityyppiToToteutuksenAjankohta = _.cond([
   [
